Klaipeda-oro-salygos.sh: Capture the direction using `read`.
Instead of going `head` on the two-line long input & then stripping spaces using `sed`, let's read the var in curly braced block (different scope).

This also introduces some irrelevant, albeit possibly annoying change.  Output of the original script:

    $ ./Klaipeda-oro-salygos.sh
    109.70°/5.43kt

Notice the degree sign put just after the direction value.
Now with this change all the symbols except leading spaces are put into the var (including the space between the direction value and a degree symbol), and so it is passed into our output:

    $ ./Klaipeda-oro-salygos.sh
    128.69 °/5.82kt

As we are going to eliminate this symbol altogether, lets keep this irrelevant regression in place.